api/requirements.txt
flask
flask-cors
paho-mqtt
flasgger
